# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlowTrack API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  skills?: string[];
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Profile {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  skills?: string[];
  /** @nullable */
  cvFilename?: string | null;
  createdAt: string;
}

export interface ProfileUpdate {
  name?: string;
  bio?: string;
  skills?: string[];
  avatar?: string;
}

export interface Job {
  id: string;
  company: string;
  role: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  salary?: string | null;
  /** @nullable */
  jobUrl?: string | null;
  appliedAt: string;
  createdAt: string;
}

export interface JobInput {
  company: string;
  role: string;
  status: string;
  notes?: string;
  salary?: string;
  jobUrl?: string;
  appliedAt?: string;
}

export interface JobUpdate {
  company?: string;
  role?: string;
  status?: string;
  notes?: string;
  salary?: string;
  jobUrl?: string;
  appliedAt?: string;
}

export interface WeeklyDataPoint {
  week: string;
  count: number;
}

export interface JobStats {
  total: number;
  applied: number;
  interview: number;
  offer: number;
  rejected: number;
  weeklyData: WeeklyDataPoint[];
}

export interface Task {
  id: string;
  title: string;
  completed: boolean;
  priority: string;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export interface TaskInput {
  title: string;
  priority: string;
  dueDate?: string;
}

export interface TaskUpdate {
  title?: string;
  completed?: boolean;
  priority?: string;
  dueDate?: string;
}

export interface Habit {
  id: string;
  name: string;
  completed: boolean;
  streak: number;
  /** @nullable */
  lastCompletedAt?: string | null;
  createdAt: string;
}

export interface HabitInput {
  name: string;
}

export interface HabitUpdate {
  name?: string;
}

export interface DashboardStats {
  totalJobs: number;
  interviews: number;
  tasksCompleted: number;
  habitStreak: number;
  todayTasksTotal: number;
  todayTasksDone: number;
}

export interface StatusBreakdown {
  status: string;
  count: number;
}

export interface DashboardActivity {
  recentJobs: Job[];
  todayTasks: Task[];
  weeklyApplications: WeeklyDataPoint[];
  jobStatusBreakdown: StatusBreakdown[];
}

export type GetJobsParams = {
status?: string;
search?: string;
};

export type GetTasksParams = {
status?: string;
};