# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GlowTrack API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: jobs
    description: Job applications
  - name: tasks
    description: Tasks
  - name: habits
    description: Habits
  - name: profile
    description: User profile
  - name: dashboard
    description: Dashboard stats

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: List all jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content

  /jobs/stats/summary:
    get:
      operationId: getJobStats
      tags: [jobs]
      summary: Get job application stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStats"

  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content

  /habits:
    get:
      operationId: getHabits
      tags: [habits]
      summary: List all habits
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"

  /habits/{id}:
    patch:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content

  /habits/{id}/toggle:
    post:
      operationId: toggleHabit
      tags: [habits]
      summary: Toggle habit completion for today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard overview stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent activity and weekly chart data
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardActivity"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        createdAt:
          type: string

    Profile:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        cvFilename:
          type: ["string", "null"]
        createdAt:
          type: string

    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        avatar:
          type: string

    Job:
      type: object
      required: [id, company, role, status, appliedAt, createdAt]
      properties:
        id:
          type: string
        company:
          type: string
        role:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        jobUrl:
          type: ["string", "null"]
        appliedAt:
          type: string
        createdAt:
          type: string

    JobInput:
      type: object
      required: [company, role, status]
      properties:
        company:
          type: string
        role:
          type: string
        status:
          type: string
        notes:
          type: string
        salary:
          type: string
        jobUrl:
          type: string
        appliedAt:
          type: string

    JobUpdate:
      type: object
      properties:
        company:
          type: string
        role:
          type: string
        status:
          type: string
        notes:
          type: string
        salary:
          type: string
        jobUrl:
          type: string
        appliedAt:
          type: string

    JobStats:
      type: object
      required: [total, applied, interview, offer, rejected, weeklyData]
      properties:
        total:
          type: integer
        applied:
          type: integer
        interview:
          type: integer
        offer:
          type: integer
        rejected:
          type: integer
        weeklyData:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyDataPoint"

    WeeklyDataPoint:
      type: object
      required: [week, count]
      properties:
        week:
          type: string
        count:
          type: integer

    Task:
      type: object
      required: [id, title, completed, priority, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        completed:
          type: boolean
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string

    TaskInput:
      type: object
      required: [title, priority]
      properties:
        title:
          type: string
        priority:
          type: string
        dueDate:
          type: string

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean
        priority:
          type: string
        dueDate:
          type: string

    Habit:
      type: object
      required: [id, name, completed, streak, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        completed:
          type: boolean
        streak:
          type: integer
        lastCompletedAt:
          type: ["string", "null"]
        createdAt:
          type: string

    HabitInput:
      type: object
      required: [name]
      properties:
        name:
          type: string

    HabitUpdate:
      type: object
      properties:
        name:
          type: string

    DashboardStats:
      type: object
      required: [totalJobs, interviews, tasksCompleted, habitStreak, todayTasksTotal, todayTasksDone]
      properties:
        totalJobs:
          type: integer
        interviews:
          type: integer
        tasksCompleted:
          type: integer
        habitStreak:
          type: integer
        todayTasksTotal:
          type: integer
        todayTasksDone:
          type: integer

    DashboardActivity:
      type: object
      required: [recentJobs, todayTasks, weeklyApplications, jobStatusBreakdown]
      properties:
        recentJobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        todayTasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        weeklyApplications:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyDataPoint"
        jobStatusBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/StatusBreakdown"

    StatusBreakdown:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer