# lib/api-zod/src/generated/types/habit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlowTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Habit {
  id: string;
  name: string;
  completed: boolean;
  streak: number;
  /** @nullable */
  lastCompletedAt?: string | null;
  createdAt: string;
}