# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlowTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobUpdate {
  company?: string;
  role?: string;
  status?: string;
  notes?: string;
  salary?: string;
  jobUrl?: string;
  appliedAt?: string;
}