# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlowTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  name?: string;
  bio?: string;
  skills?: string[];
  avatar?: string;
}