# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlowTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: string;
  title: string;
  completed: boolean;
  priority: string;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}