# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlowTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  skills?: string[];
  createdAt: string;
}