# lib/api-zod/src/generated/types/weeklyDataPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlowTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyDataPoint {
  week: string;
  count: number;
}